Move Esc, Text and Refresh into output/.
