+++ lib/vedeu/output/all.rb
@@ -1,3 +1,5 @@
+require 'vedeu/output/esc'
+
 require 'vedeu/output/translator'
 require 'vedeu/output/background'
 require 'vedeu/output/foreground'
@@ -8,13 +10,17 @@ require 'vedeu/output/compositor'
 
 require 'vedeu/output/presentation'
 require 'vedeu/output/style'
+require 'vedeu/output/text'
 require 'vedeu/output/position_index'
 require 'vedeu/output/index_position'
 require 'vedeu/output/virtual_buffer'
 require 'vedeu/output/html_char'
 require 'vedeu/output/html_renderer'
+require 'vedeu/output/refresh'
+
 require 'vedeu/output/renderer'
 require 'vedeu/output/output'
+
 require 'vedeu/output/viewport'
 require 'vedeu/output/virtual_terminal'
 require 'vedeu/output/wordwrap'
