+++ lib/vedeu/output/refresh.rb
