+++ lib/vedeu/output/text.rb
