+++ lib/vedeu/support/all.rb
@@ -1,14 +1,11 @@
 require 'vedeu/support/coercions'
 require 'vedeu/support/common'
 require 'vedeu/support/console'
-require 'vedeu/support/esc'
 require 'vedeu/support/log'
 require 'vedeu/support/read'
-require 'vedeu/support/refresh'
 require 'vedeu/support/sentence'
 require 'vedeu/support/template'
 require 'vedeu/support/terminal'
-require 'vedeu/support/text'
 require 'vedeu/support/trace'
 require 'vedeu/support/visible'
 require 'vedeu/support/write'
