+++ test/lib/vedeu/output/refresh_test.rb
