+++ test/lib/vedeu/output/text_test.rb
